Верная навигационная логика
